Upgrade Stripe's to_dict_recursive by handling more types
Looking in the SDK, `to_dict_recursive()` casts the object to a
dictionary. This seems to be broken with certain versions of the API
because we can have a StripeObject with a field that is a list of other
StripeObjects. This new function looks for StripeObjects in lists and
dictionaries and returns everything else unchanged.