Fix customer event transform error (#53)
* Upgrade Stripe's to_dict_recursive by handling more types

Looking in the SDK, `to_dict_recursive()` casts the object to a
dictionary. This seems to be broken with certain versions of the API
because we can have a StripeObject with a field that is a list of other
StripeObjects. This new function looks for StripeObjects in lists and
dictionaries and returns everything else unchanged.

* Remove old schema because it's not always valid

* Make old schema a case in an 'anyOf'

* Neww anyOf case based on Stripe docs

* New anyOf case based on transform errors

* Add a test for the 4 paths in recursive_to_dict

* Call function on correct object

Co-authored-by: Jacob Baca <52418765+jacobrobertbaca@users.noreply.github.com>